Converted All over to Single PowerShell
This is the first stop - All bginfo info is now pulled via several
PowerShell scripts. There are Get- scripts and Format- scripts. The gets
get and the formats format. All are wrapped into a handy
Get-SystemReport script that dynamically grows the report as new Gets
and Formats are added... plus it is Standalone and will soon move into a
separate repository!! Running 1 huge script is quicker (and only causes
1 console popup) - huge win! Bigly!!